refactor: changing the structure and adding configuration classes
